fix: java.util.Collections doesn't work when treating Maps as Records
java.util.Collections methods which take a Map as a parameter don't work if we just treat all Maps as Records (because the type fo the key may not be compatible with a Record). For now, just remove it from the default generated types; ideally we'd need to handle the Map type in a smarter way.